// jurisdictions/typechain-types/DeltaTransformer.sol/DeltaTransformer.ts
export declare namespace DeltaTransformer {
      
    export type ArgumentsStruct = {fillRatios: BigNumberish[], secrets: BytesLike[]}

    export type ArgumentsStructOutput = [fillRatios: bigint[], secrets: string[]] & {fillRatios: bigint[], secrets: string[] }
  

    export type PaymentStruct = {deltaIndex: BigNumberish, amount: BigNumberish, revealedUntilTimestamp: BigNumberish, hash: BytesLike}

    export type PaymentStructOutput = [deltaIndex: bigint, amount: bigint, revealedUntilTimestamp: bigint, hash: string] & {deltaIndex: bigint, amount: bigint, revealedUntilTimestamp: bigint, hash: string }
  

    export type SwapStruct = {ownerIsLeft: boolean, addDeltaIndex: BigNumberish, addAmount: BigNumberish, subDeltaIndex: BigNumberish, subAmount: BigNumberish}

    export type SwapStructOutput = [ownerIsLeft: boolean, addDeltaIndex: bigint, addAmount: bigint, subDeltaIndex: bigint, subAmount: bigint] & {ownerIsLeft: boolean, addDeltaIndex: bigint, addAmount: bigint, subDeltaIndex: bigint, subAmount: bigint }
  

    export type PullStruct = {deltaIndex: BigNumberish, amount: BigNumberish, claimedRatio: BigNumberish, fullHash: BytesLike, partialRoot: BytesLike, targetRole: boolean}

    export type PullStructOutput = [deltaIndex: bigint, amount: bigint, claimedRatio: bigint, fullHash: string, partialRoot: string, targetRole: boolean] & {deltaIndex: bigint, amount: bigint, claimedRatio: bigint, fullHash: string, partialRoot: string, targetRole: boolean }
  

    export type BatchStruct = {payment: DeltaTransformer.PaymentStruct[], swap: DeltaTransformer.SwapStruct[], pull: DeltaTransformer.PullStruct[]}

    export type BatchStructOutput = [payment: DeltaTransformer.PaymentStructOutput[], swap: DeltaTransformer.SwapStructOutput[], pull: DeltaTransformer.PullStructOutput[]] & {payment: DeltaTransformer.PaymentStructOutput[], swap: DeltaTransformer.SwapStructOutput[], pull: DeltaTransformer.PullStructOutput[] }
  
    }

  export interface DeltaTransformerInterface extends Interface {
    getFunction(nameOrSignature: "applyBatch" | "containsPull" | "decodeArgumentsStrict" | "decodeTransformerArgumentListStrict" | "encodeBatch" | "hashToTimestamp" | "revealSecret"): FunctionFragment;

    getEvent(nameOrSignatureOrTopic: "SecretRevealed"): EventFragment;

    encodeFunctionData(functionFragment: 'applyBatch', values: [BigNumberish[], BigNumberish[], BytesLike, BytesLike, BytesLike, BigNumberish, BigNumberish, BytesLike, BytesLike, BigNumberish, BigNumberish, BigNumberish, BigNumberish]): string;
encodeFunctionData(functionFragment: 'containsPull', values: [BytesLike]): string;
encodeFunctionData(functionFragment: 'decodeArgumentsStrict', values: [BytesLike]): string;
encodeFunctionData(functionFragment: 'decodeTransformerArgumentListStrict', values: [BytesLike]): string;
encodeFunctionData(functionFragment: 'encodeBatch', values: [DeltaTransformer.BatchStruct]): string;
encodeFunctionData(functionFragment: 'hashToTimestamp', values: [BytesLike]): string;
encodeFunctionData(functionFragment: 'revealSecret', values: [BytesLike]): string;

    decodeFunctionResult(functionFragment: 'applyBatch', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'containsPull', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'decodeArgumentsStrict', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'decodeTransformerArgumentListStrict', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'encodeBatch', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'hashToTimestamp', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'revealSecret', data: BytesLike): Result;
  }

  
    export namespace SecretRevealedEvent {
      export type InputTuple = [hashlock: BytesLike, secret: BytesLike];
      export type OutputTuple = [hashlock: string, secret: string];
      export interface OutputObject {hashlock: string, secret: string };
      export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>
      export type Filter = TypedDeferredTopicFilter<Event>
      export type Log = TypedEventLog<Event>
      export type LogDescription = TypedLogDescription<Event>
    }

  

  export interface DeltaTransformer extends BaseContract {
    
    connect(runner?: ContractRunner | null): DeltaTransformer;
    waitForDeployment(): Promise<this>;

    interface: DeltaTransformerInterface;

    
  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined,
  ): Promise<Array<TypedEventLog<TCEvent>>>
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(event: TCEvent, listener: TypedListener<TCEvent>): Promise<this>
  on<TCEvent extends TypedContractEvent>(filter: TypedDeferredTopicFilter<TCEvent>, listener: TypedListener<TCEvent>): Promise<this>
  
  once<TCEvent extends TypedContractEvent>(event: TCEvent, listener: TypedListener<TCEvent>): Promise<this>
  once<TCEvent extends TypedContractEvent>(filter: TypedDeferredTopicFilter<TCEvent>, listener: TypedListener<TCEvent>): Promise<this>

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>
  removeAllListeners<TCEvent extends TypedContractEvent>(event?: TCEvent): Promise<this>


    
    
    applyBatch: TypedContractMethod<
      [deltas: BigNumberish[], tokenIds: BigNumberish[], encodedBatch: BytesLike, leftArguments: BytesLike, rightArguments: BytesLike, leftArgumentsTimestamp: BigNumberish, rightArgumentsTimestamp: BigNumberish, leftEntity: BytesLike, rightEntity: BytesLike, disputeStartTimestamp: BigNumberish, disputeTimeout: BigNumberish, leftResponseSeconds: BigNumberish, rightResponseSeconds: BigNumberish, ],
      [bigint[]],
      'view'
    >
    

    
    containsPull: TypedContractMethod<
      [encodedBatch: BytesLike, ],
      [boolean],
      'view'
    >
    

    
    decodeArgumentsStrict: TypedContractMethod<
      [encoded: BytesLike, ],
      [DeltaTransformer.ArgumentsStructOutput],
      'view'
    >
    

    
    decodeTransformerArgumentListStrict: TypedContractMethod<
      [encoded: BytesLike, ],
      [string[]],
      'view'
    >
    

    
    encodeBatch: TypedContractMethod<
      [b: DeltaTransformer.BatchStruct, ],
      [string],
      'view'
    >
    

    
    hashToTimestamp: TypedContractMethod<
      [arg0: BytesLike, ],
      [bigint],
      'view'
    >
    

    
    revealSecret: TypedContractMethod<
      [secret: BytesLike, ],
      [void],
      'nonpayable'
    >
    


    getFunction<T extends ContractMethod = ContractMethod>(key: string | FunctionFragment): T;

    getFunction(nameOrSignature: 'applyBatch'): TypedContractMethod<
      [deltas: BigNumberish[], tokenIds: BigNumberish[], encodedBatch: BytesLike, leftArguments: BytesLike, rightArguments: BytesLike, leftArgumentsTimestamp: BigNumberish, rightArgumentsTimestamp: BigNumberish, leftEntity: BytesLike, rightEntity: BytesLike, disputeStartTimestamp: BigNumberish, disputeTimeout: BigNumberish, leftResponseSeconds: BigNumberish, rightResponseSeconds: BigNumberish, ],
      [bigint[]],
      'view'
    >;
getFunction(nameOrSignature: 'containsPull'): TypedContractMethod<
      [encodedBatch: BytesLike, ],
      [boolean],
      'view'
    >;
getFunction(nameOrSignature: 'decodeArgumentsStrict'): TypedContractMethod<
      [encoded: BytesLike, ],
      [DeltaTransformer.ArgumentsStructOutput],
      'view'
    >;
getFunction(nameOrSignature: 'decodeTransformerArgumentListStrict'): TypedContractMethod<
      [encoded: BytesLike, ],
      [string[]],
      'view'
    >;
getFunction(nameOrSignature: 'encodeBatch'): TypedContractMethod<
      [b: DeltaTransformer.BatchStruct, ],
      [string],
      'view'
    >;
getFunction(nameOrSignature: 'hashToTimestamp'): TypedContractMethod<
      [arg0: BytesLike, ],
      [bigint],
      'view'
    >;
getFunction(nameOrSignature: 'revealSecret'): TypedContractMethod<
      [secret: BytesLike, ],
      [void],
      'nonpayable'
    >;

    getEvent(key: 'SecretRevealed'): TypedContractEvent<SecretRevealedEvent.InputTuple, SecretRevealedEvent.OutputTuple, SecretRevealedEvent.OutputObject>;

    filters: {
      
      'SecretRevealed(bytes32,bytes32)': TypedContractEvent<SecretRevealedEvent.InputTuple, SecretRevealedEvent.OutputTuple, SecretRevealedEvent.OutputObject>;
      SecretRevealed: TypedContractEvent<SecretRevealedEvent.InputTuple, SecretRevealedEvent.OutputTuple, SecretRevealedEvent.OutputObject>;
    
    };
  }